Fix regex handling in highlightText
